+++ docs/changelog.rst
@@ -4,6 +4,18 @@ Changelog
 
 This page contains a summary of changes between the official cwepr releases. Only the biggest changes are listed here. A complete and detailed log of all changes is available through the `GitHub Repository Browser <https://github.com/tillbiskup/cwepr/commits/master>`_.
 
+Version 0.2.1
+=============
+
+Not yet released
+
+The following bugs have been fixed:
+
+* Import units correctly using the BES3T importer.
+* Magnettech goniometer sweep importer handles situation without info file.
+* Fix for import and gathering of metadata of both sources in WinEPR Importer.
+* Units are imported correctly from par file in WinEPR importer.
+
 
 Version 0.2.0
 =============